+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Fixed
+- In some occasions, nested repeat nodes that are relevant are removed from the record string as if they were irrelevant.
+
 [4.38.1] - 2017-08-10
 ---------------------
 ##### Fixed